Refactor __main__ into test_descriptivestats
